refactor: 提取公共函数
