+++ packages/ivy-design-wc/src/components/message/src/utils.ts
@@ -1,0 +1,2 @@
+import { ref } from 'vue'
+export const curMessageIndex = ref(0)
