+++ packages/ivy-design-wc/src/utils/index.ts
@@ -1,5 +1,13 @@
 export { getType, isArray, isFunction, isObject, isFunc } from './utils'
-export { isElement, trim, addClass, removeClass, hasClass, getElementsByTagName } from './dom'
+export {
+    isElement,
+    trim,
+    addClass,
+    removeClass,
+    hasClass,
+    getElementsByTagName,
+    setPropToCustomElement
+} from './dom'
 
 export const hasOwn = (obj: Record<string, any>, key: string) => {
     // eslint-disable-next-line no-prototype-builtins
